Split into `sea-connection`
