+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = [".", "sea-orm-macros", "sea-orm-codegen"]
+members = [".", "sea-connection", "sea-orm-macros", "sea-orm-codegen"]
 
 [package]
 name = "sea-orm"
@@ -31,11 +31,11 @@ futures-util = { version = "^0.3" }
 tracing = { version = "0.1", features = ["log"] }
 rust_decimal = { version = "^1", optional = true }
 sea-orm-macros = { version = "^0.7.0", path = "sea-orm-macros", optional = true }
+sea-connection = { version = "^0.7.1", path = "sea-connection" }
 sea-query = { version = "^0.23.0", features = ["thread-safe"] }
 sea-strum = { version = "^0.23", features = ["derive", "sea-orm"] }
 serde = { version = "^1.0", features = ["derive"] }
 serde_json = { version = "^1", optional = true }
-sqlx = { version = "^0.5", optional = true }
 uuid = { version = "0.8", features = ["serde", "v4"], optional = true }
 ouroboros = "0.15"
 url = "^2.2"
@@ -54,7 +54,7 @@ sea-orm = { path = ".", features = ["mock", "debug-print"] }
 pretty_assertions = { version = "^0.7" }
 
 [features]
-debug-print = []
+debug-print = ["sea-connection/debug-print"]
 default = [
     "macros",
     "with-json",
@@ -64,34 +64,28 @@ default = [
     "with-time",
 ]
 macros = ["sea-orm-macros"]
-mock = []
-with-json = ["serde_json", "sea-query/with-json", "chrono/serde"]
-with-chrono = ["chrono", "sea-query/with-chrono"]
-with-rust_decimal = ["rust_decimal", "sea-query/with-rust_decimal"]
-with-uuid = ["uuid", "sea-query/with-uuid"]
-with-time = ["time", "sea-query/with-time"]
-sqlx-all = ["sqlx-mysql", "sqlx-postgres", "sqlx-sqlite"]
-sqlx-dep = ["sqlx-json", "sqlx-chrono", "sqlx-decimal", "sqlx-uuid", "sqlx-time"]
-sqlx-json = ["sqlx/json", "with-json"]
-sqlx-chrono = ["sqlx/chrono", "with-chrono"]
-sqlx-decimal = ["sqlx/decimal", "with-rust_decimal"]
-sqlx-uuid = ["sqlx/uuid", "with-uuid"]
-sqlx-time = ["sqlx/time", "with-time"]
-sqlx-mysql = ["sqlx-dep", "sea-query/sqlx-mysql", "sqlx/mysql"]
-sqlx-postgres = ["sqlx-dep", "sea-query/sqlx-postgres", "sqlx/postgres"]
-sqlx-sqlite = ["sqlx-dep", "sea-query/sqlx-sqlite", "sqlx/sqlite"]
-runtime-async-std = []
-runtime-async-std-native-tls = [
-    "sqlx/runtime-async-std-native-tls",
-    "runtime-async-std",
-]
-runtime-async-std-rustls = [
-    "sqlx/runtime-async-std-rustls",
-    "runtime-async-std",
-]
-runtime-actix = []
-runtime-actix-native-tls = ["sqlx/runtime-actix-native-tls", "runtime-actix"]
-runtime-actix-rustls = ["sqlx/runtime-actix-rustls", "runtime-actix"]
-runtime-tokio = []
-runtime-tokio-native-tls = ["sqlx/runtime-tokio-native-tls", "runtime-tokio"]
-runtime-tokio-rustls = ["sqlx/runtime-tokio-rustls", "runtime-tokio"]
+mock = ["sea-connection/mock"]
+with-json = ["serde_json", "sea-connection/with-json"]
+with-chrono = ["chrono", "sea-connection/with-chrono"]
+with-rust_decimal = ["rust_decimal", "sea-connection/with-rust_decimal"]
+with-uuid = ["uuid", "sea-connection/with-uuid"]
+with-time = ["time", "sea-connection/with-time"]
+sqlx-all = ["sea-connection/sqlx-all"]
+sqlx-dep = ["sea-connection/sqlx-dep"]
+sqlx-json = ["sea-connection/sqlx-json"]
+sqlx-chrono = ["sea-connection/sqlx-chrono"]
+sqlx-decimal = ["sea-connection/sqlx-decimal"]
+sqlx-uuid = ["sea-connection/sqlx-uuid"]
+sqlx-time = ["sea-connection/sqlx-time"]
+sqlx-mysql = ["sea-connection/sqlx-mysql"]
+sqlx-postgres = ["sea-connection/sqlx-postgres"]
+sqlx-sqlite = ["sea-connection/sqlx-sqlite"]
+runtime-async-std =["sea-connection/runtime-async-std"]
+runtime-async-std-native-tls = ["sea-connection/runtime-async-std-native-tls"]
+runtime-async-std-rustls = ["sea-connection/runtime-async-std-rustls"]
+runtime-actix =["sea-connection/runtime-actix"]
+runtime-actix-native-tls = ["sea-connection/runtime-actix-native-tls"]
+runtime-actix-rustls = ["sea-connection/runtime-actix-rustls"]
+runtime-tokio =["sea-connection/runtime-tokio"]
+runtime-tokio-native-tls = ["sea-connection/runtime-tokio-native-tls"]
+runtime-tokio-rustls = ["sea-connection/runtime-tokio-rustls"]
