+++ sea-connection/README.md
@@ -1,0 +1,3 @@
+# SeaConnection
+
+Database Connection Layer for SeaQL
