+++ sea-connection/src/error.rs
