+++ src/entity/active_model.rs
@@ -831,261 +831,261 @@ where
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use crate::{entity::*, tests_cfg::*, DbErr};
-    use pretty_assertions::assert_eq;
+// #[cfg(test)]
+// mod tests {
+//     use crate::{entity::*, tests_cfg::*, DbErr};
+//     use pretty_assertions::assert_eq;
 
-    #[cfg(feature = "with-json")]
-    use serde_json::json;
+//     #[cfg(feature = "with-json")]
+//     use serde_json::json;
 
-    #[test]
-    #[cfg(feature = "macros")]
-    fn test_derive_into_active_model_1() {
-        mod my_fruit {
-            pub use super::fruit::*;
-            use crate as sea_orm;
-            use crate::entity::prelude::*;
+//     #[test]
+//     #[cfg(feature = "macros")]
+//     fn test_derive_into_active_model_1() {
+//         mod my_fruit {
+//             pub use super::fruit::*;
+//             use crate as sea_orm;
+//             use crate::entity::prelude::*;
 
-            #[derive(DeriveIntoActiveModel)]
-            pub struct NewFruit {
-                // id is omitted
-                pub name: String,
-                // it is required as opposed to optional in Model
-                pub cake_id: i32,
-            }
-        }
+//             #[derive(DeriveIntoActiveModel)]
+//             pub struct NewFruit {
+//                 // id is omitted
+//                 pub name: String,
+//                 // it is required as opposed to optional in Model
+//                 pub cake_id: i32,
+//             }
+//         }
 
-        assert_eq!(
-            my_fruit::NewFruit {
-                name: "Apple".to_owned(),
-                cake_id: 1,
-            }
-            .into_active_model(),
-            fruit::ActiveModel {
-                id: NotSet,
-                name: Set("Apple".to_owned()),
-                cake_id: Set(Some(1)),
-            }
-        );
-    }
+//         assert_eq!(
+//             my_fruit::NewFruit {
+//                 name: "Apple".to_owned(),
+//                 cake_id: 1,
+//             }
+//             .into_active_model(),
+//             fruit::ActiveModel {
+//                 id: NotSet,
+//                 name: Set("Apple".to_owned()),
+//                 cake_id: Set(Some(1)),
+//             }
+//         );
+//     }
 
-    #[test]
-    #[cfg(feature = "macros")]
-    fn test_derive_into_active_model_2() {
-        mod my_fruit {
-            pub use super::fruit::*;
-            use crate as sea_orm;
-            use crate::entity::prelude::*;
+//     #[test]
+//     #[cfg(feature = "macros")]
+//     fn test_derive_into_active_model_2() {
+//         mod my_fruit {
+//             pub use super::fruit::*;
+//             use crate as sea_orm;
+//             use crate::entity::prelude::*;
 
-            #[derive(DeriveIntoActiveModel)]
-            pub struct UpdateFruit {
-                pub cake_id: Option<Option<i32>>,
-            }
-        }
+//             #[derive(DeriveIntoActiveModel)]
+//             pub struct UpdateFruit {
+//                 pub cake_id: Option<Option<i32>>,
+//             }
+//         }
 
-        assert_eq!(
-            my_fruit::UpdateFruit {
-                cake_id: Some(Some(1)),
-            }
-            .into_active_model(),
-            fruit::ActiveModel {
-                id: NotSet,
-                name: NotSet,
-                cake_id: Set(Some(1)),
-            }
-        );
+//         assert_eq!(
+//             my_fruit::UpdateFruit {
+//                 cake_id: Some(Some(1)),
+//             }
+//             .into_active_model(),
+//             fruit::ActiveModel {
+//                 id: NotSet,
+//                 name: NotSet,
+//                 cake_id: Set(Some(1)),
+//             }
+//         );
 
-        assert_eq!(
-            my_fruit::UpdateFruit {
-                cake_id: Some(None),
-            }
-            .into_active_model(),
-            fruit::ActiveModel {
-                id: NotSet,
-                name: NotSet,
-                cake_id: Set(None),
-            }
-        );
+//         assert_eq!(
+//             my_fruit::UpdateFruit {
+//                 cake_id: Some(None),
+//             }
+//             .into_active_model(),
+//             fruit::ActiveModel {
+//                 id: NotSet,
+//                 name: NotSet,
+//                 cake_id: Set(None),
+//             }
+//         );
 
-        assert_eq!(
-            my_fruit::UpdateFruit { cake_id: None }.into_active_model(),
-            fruit::ActiveModel {
-                id: NotSet,
-                name: NotSet,
-                cake_id: NotSet,
-            }
-        );
-    }
+//         assert_eq!(
+//             my_fruit::UpdateFruit { cake_id: None }.into_active_model(),
+//             fruit::ActiveModel {
+//                 id: NotSet,
+//                 name: NotSet,
+//                 cake_id: NotSet,
+//             }
+//         );
+//     }
 
-    #[test]
-    #[cfg(feature = "with-json")]
-    #[should_panic(
-        expected = r#"called `Result::unwrap()` on an `Err` value: Json("missing field `id`")"#
-    )]
-    fn test_active_model_set_from_json_1() {
-        let mut cake: cake::ActiveModel = Default::default();
+//     #[test]
+//     #[cfg(feature = "with-json")]
+//     #[should_panic(
+//         expected = r#"called `Result::unwrap()` on an `Err` value: Json("missing field `id`")"#
+//     )]
+//     fn test_active_model_set_from_json_1() {
+//         let mut cake: cake::ActiveModel = Default::default();
 
-        cake.set_from_json(json!({
-            "name": "Apple Pie",
-        }))
-        .unwrap();
-    }
+//         cake.set_from_json(json!({
+//             "name": "Apple Pie",
+//         }))
+//         .unwrap();
+//     }
 
-    #[test]
-    #[cfg(feature = "with-json")]
-    fn test_active_model_set_from_json_2() -> Result<(), DbErr> {
-        let mut fruit: fruit::ActiveModel = Default::default();
+//     #[test]
+//     #[cfg(feature = "with-json")]
+//     fn test_active_model_set_from_json_2() -> Result<(), DbErr> {
+//         let mut fruit: fruit::ActiveModel = Default::default();
 
-        fruit.set_from_json(json!({
-            "name": "Apple",
-        }))?;
-        assert_eq!(
-            fruit,
-            fruit::ActiveModel {
-                id: ActiveValue::NotSet,
-                name: ActiveValue::Set("Apple".to_owned()),
-                cake_id: ActiveValue::NotSet,
-            }
-        );
+//         fruit.set_from_json(json!({
+//             "name": "Apple",
+//         }))?;
+//         assert_eq!(
+//             fruit,
+//             fruit::ActiveModel {
+//                 id: ActiveValue::NotSet,
+//                 name: ActiveValue::Set("Apple".to_owned()),
+//                 cake_id: ActiveValue::NotSet,
+//             }
+//         );
 
-        assert_eq!(
-            fruit::ActiveModel::from_json(json!({
-                "name": "Apple",
-            }))?,
-            fruit::ActiveModel {
-                id: ActiveValue::NotSet,
-                name: ActiveValue::Set("Apple".to_owned()),
-                cake_id: ActiveValue::NotSet,
-            }
-        );
+//         assert_eq!(
+//             fruit::ActiveModel::from_json(json!({
+//                 "name": "Apple",
+//             }))?,
+//             fruit::ActiveModel {
+//                 id: ActiveValue::NotSet,
+//                 name: ActiveValue::Set("Apple".to_owned()),
+//                 cake_id: ActiveValue::NotSet,
+//             }
+//         );
 
-        fruit.set_from_json(json!({
-            "name": "Apple",
-            "cake_id": null,
-        }))?;
-        assert_eq!(
-            fruit,
-            fruit::ActiveModel {
-                id: ActiveValue::NotSet,
-                name: ActiveValue::Set("Apple".to_owned()),
-                cake_id: ActiveValue::Set(None),
-            }
-        );
+//         fruit.set_from_json(json!({
+//             "name": "Apple",
+//             "cake_id": null,
+//         }))?;
+//         assert_eq!(
+//             fruit,
+//             fruit::ActiveModel {
+//                 id: ActiveValue::NotSet,
+//                 name: ActiveValue::Set("Apple".to_owned()),
+//                 cake_id: ActiveValue::Set(None),
+//             }
+//         );
 
-        fruit.set_from_json(json!({
-            "id": null,
-            "name": "Apple",
-            "cake_id": 1,
-        }))?;
-        assert_eq!(
-            fruit,
-            fruit::ActiveModel {
-                id: ActiveValue::NotSet,
-                name: ActiveValue::Set("Apple".to_owned()),
-                cake_id: ActiveValue::Set(Some(1)),
-            }
-        );
+//         fruit.set_from_json(json!({
+//             "id": null,
+//             "name": "Apple",
+//             "cake_id": 1,
+//         }))?;
+//         assert_eq!(
+//             fruit,
+//             fruit::ActiveModel {
+//                 id: ActiveValue::NotSet,
+//                 name: ActiveValue::Set("Apple".to_owned()),
+//                 cake_id: ActiveValue::Set(Some(1)),
+//             }
+//         );
 
-        fruit.set_from_json(json!({
-            "id": 2,
-            "name": "Apple",
-            "cake_id": 1,
-        }))?;
-        assert_eq!(
-            fruit,
-            fruit::ActiveModel {
-                id: ActiveValue::NotSet,
-                name: ActiveValue::Set("Apple".to_owned()),
-                cake_id: ActiveValue::Set(Some(1)),
-            }
-        );
+//         fruit.set_from_json(json!({
+//             "id": 2,
+//             "name": "Apple",
+//             "cake_id": 1,
+//         }))?;
+//         assert_eq!(
+//             fruit,
+//             fruit::ActiveModel {
+//                 id: ActiveValue::NotSet,
+//                 name: ActiveValue::Set("Apple".to_owned()),
+//                 cake_id: ActiveValue::Set(Some(1)),
+//             }
+//         );
 
-        let mut fruit = fruit::ActiveModel {
-            id: ActiveValue::Set(1),
-            name: ActiveValue::NotSet,
-            cake_id: ActiveValue::NotSet,
-        };
-        fruit.set_from_json(json!({
-            "id": 8,
-            "name": "Apple",
-            "cake_id": 1,
-        }))?;
-        assert_eq!(
-            fruit,
-            fruit::ActiveModel {
-                id: ActiveValue::Set(1),
-                name: ActiveValue::Set("Apple".to_owned()),
-                cake_id: ActiveValue::Set(Some(1)),
-            }
-        );
+//         let mut fruit = fruit::ActiveModel {
+//             id: ActiveValue::Set(1),
+//             name: ActiveValue::NotSet,
+//             cake_id: ActiveValue::NotSet,
+//         };
+//         fruit.set_from_json(json!({
+//             "id": 8,
+//             "name": "Apple",
+//             "cake_id": 1,
+//         }))?;
+//         assert_eq!(
+//             fruit,
+//             fruit::ActiveModel {
+//                 id: ActiveValue::Set(1),
+//                 name: ActiveValue::Set("Apple".to_owned()),
+//                 cake_id: ActiveValue::Set(Some(1)),
+//             }
+//         );
 
-        Ok(())
-    }
+//         Ok(())
+//     }
 
-    #[smol_potat::test]
-    #[cfg(feature = "with-json")]
-    async fn test_active_model_set_from_json_3() -> Result<(), DbErr> {
-        use crate::*;
+//     #[smol_potat::test]
+//     #[cfg(feature = "with-json")]
+//     async fn test_active_model_set_from_json_3() -> Result<(), DbErr> {
+//         use crate::*;
 
-        let db = MockDatabase::new(DbBackend::Postgres)
-            .append_exec_results(vec![
-                MockExecResult {
-                    last_insert_id: 1,
-                    rows_affected: 1,
-                },
-                MockExecResult {
-                    last_insert_id: 1,
-                    rows_affected: 1,
-                },
-            ])
-            .append_query_results(vec![
-                vec![fruit::Model {
-                    id: 1,
-                    name: "Apple".to_owned(),
-                    cake_id: None,
-                }],
-                vec![fruit::Model {
-                    id: 2,
-                    name: "Orange".to_owned(),
-                    cake_id: Some(1),
-                }],
-            ])
-            .into_connection();
+//         let db = MockDatabase::new(DbBackend::Postgres)
+//             .append_exec_results(vec![
+//                 MockExecResult {
+//                     last_insert_id: 1,
+//                     rows_affected: 1,
+//                 },
+//                 MockExecResult {
+//                     last_insert_id: 1,
+//                     rows_affected: 1,
+//                 },
+//             ])
+//             .append_query_results(vec![
+//                 vec![fruit::Model {
+//                     id: 1,
+//                     name: "Apple".to_owned(),
+//                     cake_id: None,
+//                 }],
+//                 vec![fruit::Model {
+//                     id: 2,
+//                     name: "Orange".to_owned(),
+//                     cake_id: Some(1),
+//                 }],
+//             ])
+//             .into_connection();
 
-        let mut fruit: fruit::ActiveModel = Default::default();
-        fruit.set_from_json(json!({
-            "name": "Apple",
-        }))?;
-        fruit.save(&db).await?;
+//         let mut fruit: fruit::ActiveModel = Default::default();
+//         fruit.set_from_json(json!({
+//             "name": "Apple",
+//         }))?;
+//         fruit.save(&db).await?;
 
-        let mut fruit = fruit::ActiveModel {
-            id: Set(2),
-            ..Default::default()
-        };
-        fruit.set_from_json(json!({
-            "id": 9,
-            "name": "Orange",
-            "cake_id": 1,
-        }))?;
-        fruit.save(&db).await?;
+//         let mut fruit = fruit::ActiveModel {
+//             id: Set(2),
+//             ..Default::default()
+//         };
+//         fruit.set_from_json(json!({
+//             "id": 9,
+//             "name": "Orange",
+//             "cake_id": 1,
+//         }))?;
+//         fruit.save(&db).await?;
 
-        assert_eq!(
-            db.into_transaction_log(),
-            vec![
-                Transaction::from_sql_and_values(
-                    DbBackend::Postgres,
-                    r#"INSERT INTO "fruit" ("name") VALUES ($1) RETURNING "id", "name", "cake_id""#,
-                    vec!["Apple".into()]
-                ),
-                Transaction::from_sql_and_values(
-                    DbBackend::Postgres,
-                    r#"UPDATE "fruit" SET "name" = $1, "cake_id" = $2 WHERE "fruit"."id" = $3 RETURNING "id", "name", "cake_id""#,
-                    vec!["Orange".into(), 1i32.into(), 2i32.into()]
-                ),
-            ]
-        );
+//         assert_eq!(
+//             db.into_transaction_log(),
+//             vec![
+//                 Transaction::from_sql_and_values(
+//                     DbBackend::Postgres,
+//                     r#"INSERT INTO "fruit" ("name") VALUES ($1) RETURNING "id", "name", "cake_id""#,
+//                     vec!["Apple".into()]
+//                 ),
+//                 Transaction::from_sql_and_values(
+//                     DbBackend::Postgres,
+//                     r#"UPDATE "fruit" SET "name" = $1, "cake_id" = $2 WHERE "fruit"."id" = $3 RETURNING "id", "name", "cake_id""#,
+//                     vec!["Orange".into(), 1i32.into(), 2i32.into()]
+//                 ),
+//             ]
+//         );
 
-        Ok(())
-    }
-}
+//         Ok(())
+//     }
+// }
