+++ src/executor/mod.rs
@@ -1,15 +1,13 @@
 mod delete;
-mod execute;
 mod insert;
 mod paginator;
-mod query;
 mod select;
 mod update;
 
 pub use delete::*;
-pub use execute::*;
 pub use insert::*;
 pub use paginator::*;
-pub use query::*;
 pub use select::*;
 pub use update::*;
+
+pub use sea_connection::executor::*;
